test: extract outside click tests into outside-click.test.js
